Bugfix to culling and drawing algorithms
Prevented shapes from being compared against themselves in the culling algorithm.
In the drawing algorithm if the rectangle border for a shape extended outside the extents of the grid it could have cause the evaluation function to attempt to access parts of the isFilled array which didn't exist. This has fixed that by adding checks to each bound.